requirements.txt
BeautifulSoup==3.1.0.1
Django==1.2.4
Markdown==2.0.3
MySQL-python==1.2.3
PIL==1.1.7
Sphinx==1.0.7
boto==2.0b3
django-tagging==0.3.1
easy-thumbnails==1.0-alpha-8
-e git+https://github.com/fabric/fabric.git#egg=fabric
-e git+http://github.com/django-extensions/django-extensions.git#egg=django_extensions
-e hg+http://code.welldev.org/django-storages/#egg=django-storages
-e git+http://github.com/facebook/python-sdk.git#egg=facebook
-e git+https://github.com/ojii/django-admin-filters.git#egg=django-admin-filters
-e git+https://github.com/HonzaKral/django-threadedcomments.git#egg=django-threadedcomments
-e git+https://github.com/eldarion/brabeion.git#egg=brabeion
html5lib==0.90
pisa==3.0.33
python-memcached==1.45
pycrypto==2.2
reportlab==2.4
wsgiref==0.1.2
yolk==0.4.1
vobject==0.8.1c
-e git+https://github.com/aljosa/django-tinymce.git#egg=django-tinymce
pyPdf==1.13
django-celery
django-kombu
